apps/api/requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
python-multipart
pypdf
Pillow
opencv-python-headless
python-dotenv
huggingface_hub
